Find the algorithm where I try to remove the deleted user as a member of its boards, normally what the system does is it checks if the user was the only admin of the board, if so it do not allow the removal. So i need to tweak it in such a way that, it makes one of the would-be remaining members ( an existing user ) an admin of the board, preferably the first member with the highest role to be the admin and then try the verification again and then remove the deleting user as a member of those boards.
In upstream, due to the fact that they wanted to send out mail when the user is inviting a user to the board from the board's sidebar itself, they had the same response for both cases, first, inviting an existing user and second, inviting(creating) a newly registered by admin, user. But Since we do not allow same method to send out mail if the first case is being executed, so we cannot have the same response like for the second case. The response for the second case should stay the same i.e. "E-mail sent!" and the response for the first case needs to be "Invite sent!"