Have the system verify which of the composed initial (received) scores option is activated out of the three when the modal to edit the composed intial scores.
Have the system verify which of the composed current (start) scores option is activated out of the three when the modal to edit the composed intial scores.

Have the modals that opens to edit the composed initial (received) score be populated with that card's team member's scores and/or the aspect's scores.

Have the modals that opens to edit the composed current (start) score be populated with that card's team member's scores and/or the aspect's scores.